Move types into own file
